refactor: move cm into context_managers.py
